move extension back to root level, improved build message
